+++ appboy-component/samples/android/TestApp.XamarinAndroid/MainApplication.cs
@@ -6,7 +6,6 @@ using Android.Runtime;
 
 using Com.Braze;
 using Com.Braze.Support;
-using Firebase.Iid;
 
 namespace com.appboy.xamarinsample
 {
@@ -21,18 +20,6 @@ namespace com.appboy.xamarinsample
     {
       BrazeLogger.LogLevel = 0;
       RegisterActivityLifecycleCallbacks(new BrazeActivityLifecycleCallbackListener());
-
-      new Task(RegisterFirebasePush).Start();
-    }
-
-    void RegisterFirebasePush() {
-      try {
-        var token = FirebaseInstanceId.Instance.GetToken("901477453852", "FCM");
-        Braze.GetInstance(this).RegisteredPushToken = token;
-        Console.WriteLine("Registered Firebase push token with Braze: " + token);
-      } catch (Exception e) {
-        Console.WriteLine("Caught exception registering for Firebase push: " + e);
-      }
     }
   }
 }